store: move the useSearchLangPacks hook
This change decouples the `useSearchLangPacks` hook from the
distributions sub-module in `store/api`. This hook logically belongs
with contentSources since it is using the contentSources client to get
the correct language packs for a given locale.
