+++ src/store/api/distributions/hooks.tsx
@@ -1,19 +1,16 @@
-import { useEffect, useMemo } from 'react';
+import { useMemo } from 'react';
 
 import { simpleTargetNames } from '@/constants';
 import { ImageTypes } from '@/store/api/backend';
-import { useSearchRpmMutation } from '@/store/api/contentSources';
-import { useAppDispatch, useAppSelector } from '@/store/hooks';
+import { useAppSelector } from '@/store/hooks';
 import { selectIsOnPremise } from '@/store/slices/env';
 import {
   selectArchitecture,
   selectDistribution,
   selectImageTypes,
   selectIsImageMode,
-  selectLocaleLangpackCandidates,
-  setVerifiedLocaleLangpacks,
 } from '@/store/slices/wizard';
-import { asDistribution, isImageType } from '@/store/typeGuards';
+import { isImageType } from '@/store/typeGuards';
 import isRhel from '@/Utilities/isRhel';
 
 import { ALL_CUSTOMIZATIONS } from './constants';
@@ -244,63 +241,4 @@ export const useImageTypeCustomizationSupport = (
     isOnPremise,
     isRhel: isRhel(distro),
   });
-};
-
-const extractPackageNames = (data: { package_name?: string }[]): string[] =>
-  Array.from(
-    new Set(data.flatMap((d) => (d.package_name ? [d.package_name] : []))),
-  );
-
-export const useSearchLanguagePacks = (distroUrls: string[]) => {
-  const dispatch = useAppDispatch();
-  const distribution = useAppSelector(selectDistribution);
-  const arch = useAppSelector(selectArchitecture);
-  const candidateLangpacks = useAppSelector(selectLocaleLangpackCandidates);
-  const [searchRpms, { isLoading: isSearchLoading }] = useSearchRpmMutation();
-
-  useEffect(() => {
-    if (candidateLangpacks.length === 0) {
-      dispatch(setVerifiedLocaleLangpacks([]));
-      return;
-    }
-    if (!process.env.IS_ON_PREMISE && distroUrls.length === 0) {
-      return;
-    }
-
-    let cancelled = false;
-    const run = async () => {
-      const request = process.env.IS_ON_PREMISE
-        ? {
-            packages: candidateLangpacks,
-            architecture: arch,
-            distribution: asDistribution(distribution),
-          }
-        : { exact_names: candidateLangpacks, urls: distroUrls, limit: 500 };
-
-      try {
-        const data = await searchRpms({
-          apiContentUnitSearchRequest: request,
-        }).unwrap();
-        const verified = extractPackageNames(
-          data as { package_name?: string }[],
-        );
-        if (!cancelled) dispatch(setVerifiedLocaleLangpacks(verified));
-      } catch {
-        if (!cancelled) dispatch(setVerifiedLocaleLangpacks([]));
-      }
-    };
-    run();
-    return () => {
-      cancelled = true;
-    };
-  }, [
-    arch,
-    distribution,
-    distroUrls,
-    dispatch,
-    candidateLangpacks,
-    searchRpms,
-  ]);
-
-  return { isLoading: isSearchLoading };
 };
